+++ NEWS.md
@@ -1,8 +1,12 @@
 # guideR (development version)
 
+**New feature**
+
+* new function `plot_multiple_answers_dodge()`
+
 **Fix**
 
-* Fix in `plot_multiple_answers()` with `NA` values when `drop_na = FALSE`
+* fix in `plot_multiple_answers()` with `NA` values when `drop_na = FALSE`
 
 # guideR 0.5.0